[refactor] public 대신 module로 이름 변경
