+++ .env.local
@@ -1,2 +1,3 @@
 ELASTIC_SEARCH_URL=http://localhost:9200
 MONGODB_URL=mongodb://rainbow:rainbowA307@127.0.0.1:27017
+INDEX_NAME=themes
