+++ modules/common/elasticSearch/createIndex.js
@@ -1,24 +1,22 @@
-import esIndexObject from "./esIndexObject";
 import { Client } from "@elastic/elasticsearch";
-import { ELASTIC_SEARCH_URL } from "../tools/config";
+import esIndexObject from "./esIndexObject.js";
+import { ELASTIC_SEARCH_URL, INDEX_NAME } from "../tools/config.js";
 
 const esClient = new Client({ node: ELASTIC_SEARCH_URL });
 
-const INDEX_NAME = "themes";
-
-async function createIndex() {
+const createIndex = async () => {
   try {
     // Check if the index already exists
     const indexExists = await esClient.indices.exists({ index: INDEX_NAME });
     if (indexExists) {
-      await esClient.indices.delete({ index: indexName });
+      await esClient.indices.delete({ index: INDEX_NAME });
     }
 
     await create();
   } catch (error) {
     console.error("Error creating index:", error);
   }
-}
+};
 
 const create = async () => {
   esClient.indices.create(esIndexObject(indexName));
