+++ modules/common/tools/config.js
@@ -1,2 +1,3 @@
 export const ELASTIC_SEARCH_URL = process.env.ELASTIC_SEARCH_URL;
 export const MONGODB_URL = process.env.MONGODB_URL;
+export const INDEX_NAME = process.env.INDEX_NAME;
